# feat: reduce WriteByte to improve Compact's performance