--- a/compact.go
+++ b/compact.go
@@ -4,48 +4,78 @@ import (
 	"bytes"
 )
 
+type asciiSet [8]uint32
+
+func makeASCIISet(chars string) asciiSet {
+	var as asciiSet
+	for _, c := range chars {
+		as[c>>5] |= 1 << uint(c&31)
+	}
+	return as
+}
+
+func (as *asciiSet) Contains(c byte) bool {
+	return (as[c>>5] & (1 << uint(c&31))) != 0
+}
+
+var flushSet = makeASCIISet("\" \t\n\r")
+var escapeSet = makeASCIISet("<>&")
+
 func compact(dst *bytes.Buffer, src []byte, escape bool) error {
-	length := len(src)
-	for cursor := 0; cursor < length; cursor++ {
-		c := src[cursor]
-		switch c {
-		case ' ', '\t', '\n', '\r':
-			continue
-		case '"':
-			if err := dst.WriteByte(c); err != nil {
-				return err
-			}
-			for {
-				cursor++
-				c := src[cursor]
-				if escape && (c == '<' || c == '>' || c == '&') {
-					if _, err := dst.WriteString(`\u00`); err != nil {
-						return err
-					}
-					if _, err := dst.Write([]byte{hex[c>>4], hex[c&0xF]}); err != nil {
-						return err
-					}
-				} else if err := dst.WriteByte(c); err != nil {
-					return err
-				}
-				switch c {
-				case '\\':
-					cursor++
-					if err := dst.WriteByte(src[cursor]); err != nil {
-						return err
-					}
-				case '"':
-					goto LOOP_END
-				case nul:
-					return errUnexpectedEndOfJSON("string", int64(length))
-				}
-			}
-		default:
-			if err := dst.WriteByte(c); err != nil {
-				return err
+	start := 0
+	cursor := 0
+	for {
+		offset := -1
+		flushSet := flushSet
+		for i, c := range src[cursor:] {
+			if flushSet.Contains(c) {
+				offset = i
+				break
 			}
 		}
-	LOOP_END:
+		if offset == -1 {
+			break
+		}
+		cursor += offset
+		c := src[cursor]
+		if c == '"' { // string literal
+			cursor++
+			escapeSet := escapeSet
+			for escaped := false; cursor < len(src); cursor++ {
+				c := src[cursor]
+				if !escaped {
+					if c == '"' {
+						break
+					}
+					if c == '\\' {
+						escaped = true
+					} else if escape && escapeSet.Contains(c) {
+						if start < cursor {
+							dst.Write(src[start:cursor])
+						}
+						start = cursor + 1
+						dst.Write([]byte{'\\', 'u', '0', '0', hex[c>>4], hex[c&0xf]})
+					}
+				} else {
+					escaped = false
+					if escape && escapeSet.Contains(c) {
+						return errInvalidCharacter(c, "escaped string", int64(cursor))
+					}
+				}
+			}
+			if cursor == len(src) {
+				return errUnexpectedEndOfJSON("string", int64(len(src)))
+			}
+		} else { // whitespaces
+			if start < cursor {
+				dst.Write(src[start:cursor])
+			}
+			start = cursor + 1
+		}
+		cursor++
+	}
+	if start < len(src) {
+		dst.Write(src[start:])
 	}
 	return nil
 }