I reduced WriteByte to improve Compact's performance.

## go-json benchmarks
```
name                          old time/op    new time/op    delta
_Encode_MarshalJSON_GoJson-8     100ns ± 2%      92ns ± 1%  -8.17%  (p=0.000 n=10+10)

name                          old alloc/op   new alloc/op   delta
_Encode_MarshalJSON_GoJson-8     16.0B ± 0%     16.0B ± 0%    ~     (all equal)

name                          old allocs/op  new allocs/op  delta
_Encode_MarshalJSON_GoJson-8      2.00 ± 0%      2.00 ± 0%    ~     (all equal)
```

## lestrrat-go/jwx benchmarks
```
name                                                    old time/op    new time/op    delta
JWE/Serialization/JSON/json.Marshal                       13.4µs ± 2%    13.2µs ± 1%   -1.47%  (p=0.006 n=9+9)
JWK/Serialization/RSA/PublicKey/json.Marshal              5.97µs ± 2%    5.06µs ± 2%  -15.26%  (p=0.000 n=9+10)
JWK/Serialization/RSA/PrivateKey/json.Marshal             17.2µs ± 1%    13.5µs ± 1%  -21.49%  (p=0.000 n=10+10)
JWK/Serialization/EC/PublicKey/json.Marshal               5.56µs ± 2%    5.05µs ± 1%   -9.25%  (p=0.000 n=9+10)
JWK/Serialization/EC/PrivateKey/json.Marshal              6.59µs ± 1%    5.83µs ± 1%  -11.57%  (p=0.000 n=10+10)
JWK/Serialization/Symmetric/PublicKey/json.Marshal        3.85µs ± 1%    3.52µs ± 1%   -8.73%  (p=0.000 n=9+10)
JWK/Serialization/Symmetric/PrivateKey/json.Marshal       3.84µs ± 1%    3.51µs ± 1%   -8.61%  (p=0.000 n=10+10)
JWS/Serialization/JSON/json.Marshal                       14.2µs ± 1%    14.2µs ± 1%     ~     (p=0.870 n=10+10)
JWT/Serialization/JSON/json.Marshal                       3.00µs ± 2%    3.00µs ± 1%     ~     (p=0.504 n=9+9)
```