@orisano 2021-03-11
方向性が違うようならPR自体を閉じてもらっても全然構わないです！