@goccy 2021-03-11
ありがとうございます。ではいったん閉じさせていただこうかなと思います。
修正は参考にさせていただきます！ 🙇